Make EMCAL trigger normalization task a general framework task
- Move to PWG/EMCAL/EMCALtask
- Add basing doxygen documentation
