+++ PWG/EMCAL/macros/AddTaskEmcalTriggerNormalization.C
@@ -1,0 +1,3 @@
+PWG::EMCAL::AliAnalysisTaskEmcalTriggerNormalization *AddTaskEmcalTriggerNormalization(const char *name) {
+  return PWG::EMCAL::AliAnalysisTaskEmcalTriggerNormalization::AddTaskEmcalTriggerNormalization(name);
+}
