--- PWGJE/EMCALJetTasks/macros/AddTaskEmcalTriggerNormalization.C
@@ -1,3 +1,0 @@
-PWGJE::EMCALJetTasks::AliAnalysisTaskEmcalTriggerNormalization *AddTaskEmcalTriggerNormalization(const char *name) {
-  return PWGJE::EMCALJetTasks::AliAnalysisTaskEmcalTriggerNormalization::AddTaskEmcalTriggerNormalization(name);
-}
